Internal Practical Exam/requirements.txt
# Transfer Learning — Fruit Classifier
# Install with: pip install -r requirements.txt

tensorflow>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pillow>=10.0.0
requests>=2.31.0
jupyter>=1.0.0
ipykernel>=6.25.0
